# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MOUSEKEEPER Control Plane API
  version: 0.1.0
  description: PostgreSQL에 먼저 영속되는 MOUSEKEEPER control-plane 공개 계약
security:
  - firebaseBearer: []
paths:
...
  /v1/rooms:
    get:
      responses:
        "200": { description: Owned active rooms }
    post:
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateRoom" } }
      responses:
        "201": { description: Room metadata created }
  /v1/rooms/{roomId}:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    delete:
      security: [{ deviceBearer: [] }]
      responses:
        "200":
          {
            description: Room removed and active transfer/cache object deletion scheduled,
          }
        "404":
          {
            description: Active room not found for the authenticated desktop device,
          }
...
  /v1/rooms/{roomId}/rules:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Rules ordered by priority }
    post:
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateRule" } }
      responses:
        "201": { description: Versioned deterministic rule }
  /v1/rules/{ruleId}:
    parameters:
      - {
          in: path,
          name: ruleId,
          required: true,
          schema: { type: string, format: uuid },
        }
    patch:
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/UpdateRule" } }
      responses:
        "200": { description: Updated rule with incremented version }
        "409": { description: Optimistic version conflict }
...
  /v1/rooms/{roomId}/file-transfers:
    parameters:
      - {
          in: path,
          name: roomId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      parameters:
        - {
            in: header,
            name: Idempotency-Key,
            required: true,
            schema: { type: string, minLength: 8, maxLength: 128 },
          }
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CreateFileTransfer" } }
      responses:
        "201": { description: Expiring transfer session }
        "409": { description: Desktop is offline }
  /v1/agent/file-transfers/{transferId}/complete-upload:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
      - {
          in: header,
          name: Idempotency-Key,
          required: true,
          schema: { type: string, minLength: 8, maxLength: 128 },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/CompleteFileUpload" } }
      responses:
        "201": { description: Upload checksum and object size accepted }
  /v1/devices/{deviceId}/file-transfers/pending:
    parameters:
      - {
          in: path,
          name: deviceId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      security: [{ deviceBearer: [] }]
      responses:
        "200":
          {
            description: Pending transfer requests for the exact desktop device,
          }
  /v1/agent/file-transfers/{transferId}/upload-target:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/RequestUploadTarget" } }
      responses:
        "201":
          {
            description: Session-bounded signed upload target after source validation,
          }
        "503": { description: Object storage is not configured }
  /v1/agent/file-transfers/{transferId}/failure:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      security: [{ deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            { schema: { $ref: "#/components/schemas/FailFileTransfer" } }
      responses:
        "201":
          {
            description: Source validation failure persisted and partial object cleanup scheduled,
          }
  /v1/file-transfers/{transferId}:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200": { description: Transfer state }
    delete:
      responses:
        "200": { description: Transfer cancelled and object deletion scheduled }
  /v1/file-transfers/{transferId}/download:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
    get:
      responses:
        "200":
          { description: At most 60-second and session-bounded download target }
  /v1/file-transfers/{transferId}/ack:
    parameters:
      - {
          in: path,
          name: transferId,
          required: true,
          schema: { type: string, format: uuid },
        }
    post:
      responses:
        "201":
          {
            description: Mobile checksum verification acknowledged and deletion scheduled,
          }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    deviceBearer:
      type: http
      scheme: bearer
      bearerFormat: mk_device_ JWT
  schemas:
...
    CreateRoom:
      type: object
      additionalProperties: false
      required: [desktopDeviceId, name, rootAlias]
      properties:
        desktopDeviceId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        rootAlias: { type: string, minLength: 1, maxLength: 120 }
...
    CreateFileTransfer:
      type: object
      additionalProperties: false
      required: [sourceRelativePath]
      properties:
        sourceRelativePath: { type: string, minLength: 1, maxLength: 1024 }
    CompleteFileUpload:
      type: object
      additionalProperties: false
      required: [sizeBytes, sha256]
      properties:
        sizeBytes: { type: integer, minimum: 1 }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
    RequestUploadTarget:
      type: object
      additionalProperties: false
      required: [sourceVersion]
      properties:
        sourceVersion:
          type: object
          additionalProperties: false
          required: [fileId, sizeBytes, modifiedAt]
          properties:
            fileId: { type: string, minLength: 1, maxLength: 512 }
            sizeBytes: { type: integer, minimum: 1 }
            modifiedAt: { type: string, format: date-time }
    FailFileTransfer:
      type: object
      additionalProperties: false
      required: [failureCode]
      properties:
        failureCode:
          {
            type: string,
            enum:
              [
                SOURCE_NOT_FOUND,
                SOURCE_CHANGED,
                OUTSIDE_MANAGED_ROOT,
                SIZE_LIMIT_EXCEEDED,
                CHECKSUM_MISMATCH,
              ],
          }
...
    RuleDefinition:
      type: object
      additionalProperties: false
      required: [conditions, action]
      properties:
        match: { type: string, enum: [ALL, ANY], default: ALL }
        conditions:
          type: array
          minItems: 1
          maxItems: 10
          items:
            oneOf:
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: extension }
                  operator: { const: IN }
                  value:
                    {
                      type: array,
                      minItems: 1,
                      maxItems: 50,
                      items: { type: string, pattern: '^\\.[a-zA-Z0-9]+$' },
                    }
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: ageDays }
                  operator: { const: GTE }
                  value: { type: integer, minimum: 0, maximum: 36500 }
              - type: object
                additionalProperties: false
                required: [field, operator, value]
                properties:
                  field: { const: name }
                  operator:
                    { type: string, enum: [CONTAINS, STARTS_WITH, ENDS_WITH] }
                  value: { type: string, minLength: 1, maxLength: 120 }
        action:
          oneOf:
            - type: object
              additionalProperties: false
              required: [type, destinationTemplate]
              properties:
                type: { const: MOVE }
                destinationTemplate: { type: string, minLength: 1 }
            - type: object
              additionalProperties: false
              required: [type]
              properties:
                type: { const: QUARANTINE }
    CreateRule:
      type: object
      additionalProperties: false
      required: [name, definition, priority]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        definition: { $ref: "#/components/schemas/RuleDefinition" }
        priority: { type: integer, minimum: 0, maximum: 10000 }
        enabled: { type: boolean, default: true }
    UpdateRule:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
        name: { type: string, minLength: 1, maxLength: 120 }
        definition: { $ref: "#/components/schemas/RuleDefinition" }
        priority: { type: integer, minimum: 0, maximum: 10000 }
        enabled: { type: boolean }
...